src/requirements.txt
# Core Dependencies
requests==2.31.0
boto3==1.35.0
openai==1.54.3
python-dotenv==1.0.0
pytz==2024.2
httpx==0.27.2

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7  # For JWT signature verification

# Data Validation
pydantic==2.5.3

# AWS Lambda Runtime (if deploying to Lambda)
# aws-lambda-powertools==2.28.0

# Monitoring (optional)
# aws-xray-sdk==2.12.0
# watchtower==3.0.1  # CloudWatch logging
